total renaming gui and lib files
