--- gui/poly_obstacle.h
@@ -1,1 +1,0 @@
-#pragma once
