+++ gui/trappy_line.h
@@ -3,7 +3,7 @@
 #include <initializer_list>
 #include <vector>
 
-#include "../lib/no_direct_fly_rule.h"
+#include "../lib/trappy_line.h"
 
 // #include "base.h"
 #include "target.h"
