+++ lib/trappy_circle.h
